Use floats

+++ auen/run-auen.py
@@ -13,8 +13,8 @@ if len(sys.argv) != 4:
     print("Requires three arguments: <NI> <NE> <OUTPUT>")
     sys.exit(1)
 
-N1 = int(sys.argv[1])
-NE = int(sys.argv[2])
+N1 = float(sys.argv[1])
+NE = float(sys.argv[2])
 output = sys.argv[3]
     
 auen_ff.run_one(N1, NE, output)
